Tests/LibWeb/Text/expected/keyboard-scroll-with-keypad-modifier.txt

before: 0
after: 100
